+++ deepinv/models/drunet.py
@@ -20,6 +20,9 @@ class DRUNet(nn.Module):
 
     The network takes into account the noise level of the input image, which is encoded as an additional input channel.
 
+    A pretrained network for (in_channels=out_channels=1 or in_channels=out_channels=3)
+    can be downloaded via setting ``pretrained='download'``.
+
     :param int in_channels: number of channels of the input.
     :param int out_channels: number of channels of the output.
     :param list nc: number of convolutional layers.
@@ -32,7 +35,7 @@ class DRUNet(nn.Module):
         shuffling, and "upconv" for nearest neighbour upsampling with additional convolution.
     :param str, None pretrained: use a pretrained network. If ``pretrained=None``, the weights will be initialized at random
         using Pytorch's default initialization. If ``pretrained='download'``, the weights will be downloaded from an
-        online repository (only available for the default architecture).
+        online repository (only available for the default architecture with 3 or 1 input/output channels).
         Finally, ``pretrained`` can also be set as a path to the user's own pretrained weights.
     :param bool train: training or testing mode.
     :param str device: gpu or cpu.
