+++ deepinv/sampling/__init__.py
@@ -1,2 +1,2 @@
 from .langevin import MonteCarlo, ULA, SKRock
-from .diffusion import DDRM, DiffusionSampler, DiffPIR
+from .diffusion import DDRM, DiffusionSampler, DiffPIR, DPS
