+++ deepinv/utils/plotting.py
@@ -11,7 +11,6 @@ import matplotlib
 
 matplotlib.rcParams.update({"font.size": 17})
 matplotlib.rcParams["lines.linewidth"] = 2
-matplotlib.style.use("seaborn-v0_8-darkgrid")
 from matplotlib.ticker import MaxNLocator
 
 plt.rcParams["text.usetex"] = True
