+++ examples/sampling/demo_diffpir.py
@@ -4,12 +4,12 @@ Implementing DiffPIR
 
 In this tutorial, we revisit the implementation of the DiffPIR diffusion algorithm for image reconstruction from
 `Zhou et al. <https://arxiv.org/abs/2305.08995>`_. The full algorithm is implemented in
-:class:`deepinv.sampling.diffusion.DiffPIR`.
+:class:`deepinv.sampling.DiffPIR`.
 """
 import numpy as np
 import torch
 import matplotlib.pyplot as plt
-
+from tqdm import tqdm
 import deepinv as dinv
 from deepinv.utils.plotting import plot
 from deepinv.optim.data_fidelity import L2
@@ -18,8 +18,13 @@ from deepinv.optim.data_fidelity import L2
 # Generate an inverse problem
 # ---------------------------
 #
-# We first generate a deblurring problem with the Butterly image. We use a square blur kernel of size 5x5 and
+# We first generate a deblurring problem with the Butterfly image. We use a square blur kernel of size 5x5 and
 # Gaussian noise with standard deviation 12.75/255.0.
+#
+# .. note::
+#           We work with an image of size 64x64 to reduce the computational time of this example.
+#           The algorithm works best with images of size 256x256.
+#
 
 from deepinv.utils.demo import load_url_image
 
@@ -29,16 +34,16 @@ url = (
     "https://mycore.core-cloud.net/index.php/s/9EzDqcJxQUJKYul/"
     "download?path=%2Fdatasets&files=butterfly.png"
 )
-x_true = load_url_image(url=url, img_size=256)
+x_true = load_url_image(url=url, img_size=64, device=device)
 x = x_true.clone()
 
-sigma = 12.75 / 255.0  # noise level
+sigma_noise = 12.75 / 255.0  # noise level
 
 physics = dinv.physics.BlurFFT(
     img_size=(3, x.shape[-2], x.shape[-1]),
     filter=torch.ones((1, 1, 5, 5), device=device) / 25,
     device=device,
-    noise_model=dinv.physics.GaussianNoise(sigma=sigma),
+    noise_model=dinv.physics.GaussianNoise(sigma=sigma_noise),
 )
 
 y = physics(x)
@@ -46,7 +51,7 @@ y = physics(x)
 imgs = [y, x_true]
 plot(
     imgs,
-    titles=["measurement", "groundtruth"],
+    titles=["measurement", "ground-truth"],
 )
 
 # %%
@@ -60,17 +65,27 @@ plot(
 # .. math::
 #         \begin{equation*}
 #         \begin{aligned}
-#         x_{0}^{t} &= \left(x_t + (1 - \overline{\alpha}_t)\mathbf{s}_\theta(x_t,t))\right)/\sqrt{\overline{\alpha}_t} \\
-#         \widehat{x}_{0}^{t} &= \operatorname{prox}_{2 f(y, \cdot) /{\rho_t}}(x_{0}^{t}) \\
-#         \widehat{\varepsilon} &= \left(x_t - \sqrt{\overline{\alpha}_t} \,\, \widehat{x}_{0}^t\right)/\sqrt{1-\overline{\alpha}_t} \\
-#         \varepsilon_t &= \mathcal{N}(0, \mathbf{I}) \\
-#         x_{t-1} &= \sqrt{\overline{\alpha}_t} \,\, \widehat{x}_{0}^t + \sqrt{1-\overline{\alpha}_t} \left(\sqrt{1-\zeta} \,\, \widehat{\varepsilon} + \sqrt{\zeta} \,\, \varepsilon_t\right),
+#         \mathbf{x}_{0}^{t} &= \denoiser{\mathbf{x}_t}{\sqrt{1-\overline{\alpha}_t}/\sqrt{\overline{\alpha}_t}} \\
+#         \widehat{\mathbf{x}}_{0}^{t} &= \operatorname{prox}_{2 f(y, \cdot) /{\rho_t}}(\mathbf{x}_{0}^{t}) \\
+#         \widehat{\mathbf{\varepsilon}} &= \left(\mathbf{x}_t - \sqrt{\overline{\alpha}_t} \,\,
+#         \widehat{\mathbf{x}}_{0}^t\right)/\sqrt{1-\overline{\alpha}_t} \\
+#         \mathbf{\varepsilon}_t &= \mathcal{N}(0, \mathbf{I}) \\
+#         \mathbf{x}_{t-1} &= \sqrt{\overline{\alpha}_t} \,\, \widehat{\mathbf{x}}_{0}^t + \sqrt{1-\overline{\alpha}_t}
+#         \left(\sqrt{1-\zeta} \,\, \widehat{\mathbf{\varepsilon}} + \sqrt{\zeta} \,\, \mathbf{\varepsilon}_t\right),
 #         \end{aligned}
 #         \end{equation*}
 #
-# where :math:`\mathbf{s}_\theta` is a denoising network, :math:`\mathcal{N}(0, \mathbf{I})` is a Gaussian noise
-# with zero mean and unit variance, and :math:`\zeta` is a parameter that controls the amount of noise added at each
-# iteration. The parameters :math:`(\overline{\alpha}_t)_{0\leq t\leq T}` and :math:`(\rho_t)_{0\leq t\leq T}` are
+# where :math:`\denoiser{\cdot}{\sigma}` is a denoising network with noise level :math:`\sigma`,
+# :math:`\mathcal{N}(0, \mathbf{I})` is a Gaussian noise
+# with zero mean and unit variance, :math:`\zeta` is a parameter that controls the amount of noise added at each
+# iteration and :math:`f` refers to the data fidelity/measurement consistency term,
+# which for Gaussian Noise (implemented as :class:`deepinv.optim.L2`) is given by:
+#
+# .. math::
+#               f(\mathbf{y}, \mathbf{x}) = \frac{1}{2}\|\mathbf{y} - \mathcal{A}(\mathbf{x})\|^2
+#
+# Note that other data fidelity terms can be used, such as :class:`deepinv.optim.PoissonLikelihood`.
+# The parameters :math:`(\overline{\alpha}_t)_{0\leq t\leq T}` and :math:`(\rho_t)_{0\leq t\leq T}` are
 # sequences of positive numbers, which we will detail later on.
 #
 # Let us now implement each step of this algorithm.
@@ -82,85 +97,41 @@ plot(
 #
 # In this section, we show how to use the denoising diffusion model from DiffPIR.
 # The denoising step is implemented by a denoising network conditioned on the noise power. The authors
-# of DiffPIR use a U-Net architecture, which can be loaded as follows:
-from deepinv.models import get_diffpir_model_defaults
+# of DiffPIR use a U-Net architecture from `Ho et al. <https://arxiv.org/abs/2108.02938>`_,
+# which can be loaded as follows:
 
-model = get_diffpir_model_defaults()
+model = dinv.models.DiffUNet(large_model=False).to(device)
 
 # %%
-# Now, recall that the forward diffusion can be rewritten as, for all :math:`t`,
-#
-# .. math::
-#         \begin{equation*}
-#         x_t = \sqrt{\overline{\alpha}_t} \widehat{x}_{0} + \sqrt{1-\overline{\alpha}_t} \varepsilon_t.
-#         \end{equation*}
-#
-# In this context, the denoising diffusion model :math:`s_theta(x, t)` is trained to approximate :math:`\varepsilon_t`
-# from :math:`x_t` and :math:`t`. The denoised image thus writes:
-#
-# .. math::
-#         \begin{equation*}
-#         \widehat{x}_{0} \approx \frac{1}{\sqrt{\overline{\alpha}_t}}x_t-(\frac{1}{\sqrt{\overline{\alpha}_t}}-1) \mathbf{s}_\theta(x_t, t)
-#         \end{equation*}
 #
 # Before being able to use the pretrained model, we need to define the sequence
-# :math:`(\overline{\alpha}_t)_{0\leq t\leq T}`. In practice, we will not use :math:`\overline{\alpha}_t` directly,
-# but rather other sequences directly deriving from it (e.g. :math:`(1/\sqrt{\overline{\alpha}_t})_{0\leq t\leq T}`).
-# The following function returns these sequences:
+# :math:`(\overline{\alpha}_t)_{0\leq t\leq T}`.
+# The following function returns these sequence:
 #
 
+T = 1000  # Number of timesteps used during training
 
-num_train_timesteps = 1000  # Number of timesteps used during training
 
-
-def get_alpha_beta(
-    beta_start=0.1 / 1000, beta_end=20 / 1000, num_train_timesteps=num_train_timesteps
-):
+def get_alphas(beta_start=0.1 / 1000, beta_end=20 / 1000, num_train_timesteps=T):
     betas = np.linspace(beta_start, beta_end, num_train_timesteps, dtype=np.float32)
     betas = torch.from_numpy(betas).to(device)
     alphas = 1.0 - betas
     alphas_cumprod = np.cumprod(alphas.cpu(), axis=0)  # This is \overline{\alpha}_t
-
-    # Useful sequences deriving from alphas_cumprod
-    sqrt_alphas_cumprod = torch.sqrt(alphas_cumprod)
-    sqrt_1m_alphas_cumprod = torch.sqrt(1.0 - alphas_cumprod)
-    reduced_alpha_cumprod = torch.div(
-        sqrt_1m_alphas_cumprod, sqrt_alphas_cumprod
-    )  # equivalent noise sigma on image
-    sqrt_recip_alphas_cumprod = torch.sqrt(1.0 / alphas_cumprod)
-    sqrt_recipm1_alphas_cumprod = torch.sqrt(1.0 / alphas_cumprod - 1)
-
-    return (
-        sqrt_1m_alphas_cumprod,
-        reduced_alpha_cumprod,
-        sqrt_alphas_cumprod,
-        sqrt_recip_alphas_cumprod,
-        sqrt_recipm1_alphas_cumprod,
-        betas,
-    )
+    return torch.tensor(alphas_cumprod)
 
 
-(
-    sqrt_1m_alphas_cumprod,
-    reduced_alpha_cumprod,
-    sqrt_alphas_cumprod,
-    sqrt_recip_alphas_cumprod,
-    sqrt_recipm1_alphas_cumprod,
-    betas,
-) = get_alpha_beta()
+alphas = get_alphas()
 
-# plot the alpha sequence
-plt.rcParams.update({"font.size": 9})
-plt.figure(figsize=(6, 3))
-plt.plot(reduced_alpha_cumprod)
-plt.ylabel(r"$\frac{\sqrt{1-\overline{\alpha}_t}}{\sqrt{\overline{\alpha}_t}}$")
-plt.xlabel(r"$t$")
-plt.tight_layout()
-plt.show()
 
 # %%
 # Now that we have the sequence of interest, there remains to link noise power to the timestep. The following function
-# returns the timestep corresponding to a given noise power.
+# returns the timestep corresponding to a given noise power, which is given by
+#
+# .. math::
+#           \sigma_t = \sqrt{1-\overline{\alpha}_t}/\overline{\alpha}_t.
+
+
+sigmas = torch.sqrt(1.0 - alphas) / alphas.sqrt()
 
 
 def find_nearest(array, value):
@@ -169,62 +140,35 @@ def find_nearest(array, value):
     return idx
 
 
+t = 100  # choose arbitrary timestep
+
 # We can now apply the model to a noisy image. We first generate a noisy image
-x_noisy = x_true + torch.randn_like(x_true) * sigma
+x_noisy = x_true + torch.randn_like(x_true) * sigmas[t]
 
-# Next, we compute the timestep corresponding to the noise power sigma
-t_i = find_nearest(reduced_alpha_cumprod.cpu().numpy(), sigma * 2)
+den = model(x_noisy, sigmas[t])
 
-# Eventually, we apply the model to the noisy image
-x_noisy = 2 * x_noisy - 1  # Rescale to [-1, 1] (the model expects this range)
-noise_est = model(x_noisy, torch.tensor([t_i]).to(y.device))
-
-# The model outputs a tensor of shape (1, 6, H, W), where the last three channels correspond to a standard
-# deviation map. We only keep the first three channels, which correspond to the point estimate.
-noise_est = noise_est[:, :3, ...]
-
-# Compute the denoised image
-den = (
-    sqrt_recip_alphas_cumprod[t_i] * x_noisy
-    - sqrt_recipm1_alphas_cumprod[t_i] * noise_est
-)
-
-# Rescale to [0, 1]
-x_noisy = x_noisy * 0.5 + 0.5
-noise_est = noise_est * 0.5 + 0.5
-den = den * 0.5 + 0.5
-
-imgs = [x_noisy, noise_est, den, den - x_true]
+imgs = [x_noisy, den, den - x_true]
 plot(
     imgs,
-    titles=["noisy input", "model output", "denoised image", "error"],
+    titles=["noisy input", "denoised image", "error"],
 )
 
 # %%
 # Data fidelity step
 # ------------------
 #
-# The data fidelity step is easily implemented in deepinv. We simply need to define a data fidelity function and use
+# The data fidelity step is easily implemented in this library. We simply need to define a data fidelity function and use
 # its prox attribute. For instance:
 
 data_fidelity = L2()
 
 # In order to take a meaningful data fidelity step, it is best if we apply it to denoised measurements.
-# First, rescale the measurements in [-1, 1] and denoise it
-y_scaled = 2 * y - 1
-noise_est_sample_var = model(y_scaled, torch.tensor([t_i]).to(y.device))
-noise_est = noise_est_sample_var[:, :3, ...]
-y_denoised_scaled = (
-    sqrt_recip_alphas_cumprod[t_i] * y_scaled
-    - sqrt_recipm1_alphas_cumprod[t_i] * noise_est
-)
-
-# Rescale back to [0, 1]
-y_denoised = (y_denoised_scaled + 1) / 2
+# First, denoise the measurements:
+y_denoised = model(y, sigmas[t])
 
 # Next, apply the proximity operator of the data fidelity term (this is the data fidelity step). In the algorithm,
-# the regularization parameter is carefully chosen. Here, for simplicity, we set it to 1/sigma.
-x_prox = data_fidelity.prox(y_denoised, y, physics, gamma=1 / sigma)
+# the regularization parameter is carefully chosen. Here, for simplicity, we set it to :math:`1/\sigma`.
+x_prox = data_fidelity.prox(y_denoised, y, physics, gamma=1 / sigmas[t])
 
 imgs = [y, y_denoised, x_prox]
 plot(
@@ -237,26 +181,34 @@ plot(
 # -------------
 #
 # The last step to be implemented is the DiffPIR sampling step and this can be computed in two steps.
-# Firstly, we need to compute the effective noise in the estimated reconstruction, i.e. the residual between the previous
+# Firstly, we need to compute the effective noise in the estimated reconstruction,
+# i.e. the residual between the previous
 # reconstruction and the data fidelity step. This is done as follows:
+#
+# .. note::
+#           The diffPIR algorithm assumes that the images are in the range [-1, 1], whereas standard denoisers
+#           usually output images in the range [0, 1]. This is why we rescale the images before applying the steps.
 
 x_prox_scaled = 2 * x_prox - 1  # Rescale the output of the proximal step in [-1, 1]
-eps = (y_scaled - sqrt_alphas_cumprod[t_i] * x_prox_scaled) / sqrt_1m_alphas_cumprod[
-    t_i
-]
+y_scaled = 2 * y - 1  # Rescale the measurement in [-1, 1]
+
+t_i = find_nearest(
+    sigmas.cpu().numpy(), sigma_noise * 2
+)  # time step associated with the noise level sigma
+eps = (y_scaled - alphas[t_i].sqrt() * x_prox_scaled) / torch.sqrt(
+    1.0 - alphas[t_i]
+)  # effective noise
 
 # (notice the rescaling)
 #
 # Secondly, we need to perform the sampling step, which is a linear combination between the estimated noise and
 # the realizations of a Gaussian white noise. This is done as follows:
 zeta = 0.3
-x_sampled_scaled = (
-    sqrt_alphas_cumprod[t_i - 1] * x_prox_scaled
-    + np.sqrt(1 - zeta) * (torch.sqrt(sqrt_1m_alphas_cumprod[t_i - 1] ** 2) * eps)
-    + np.sqrt(zeta) * sqrt_1m_alphas_cumprod[t_i - 1] * torch.randn_like(x)
-)
+x_sampled_scaled = alphas[t_i - 1].sqrt() * x_prox_scaled + torch.sqrt(
+    1.0 - alphas[t_i - 1]
+) * (np.sqrt(1 - zeta) * eps + np.sqrt(zeta) * torch.randn_like(x))
 
-x_sampled = (x_sampled_scaled + 1) / 2
+x_sampled = (x_sampled_scaled + 1) / 2  # Rescale the output in [0, 1]
 
 imgs = [y, y_denoised, x_prox, x_sampled]
 plot(
@@ -276,48 +228,36 @@ plot(
 # We can now put all the steps together and implement the DiffPIR algorithm. The only remaining step is to set the
 # noise schedule (i.e. the sequence of noise powers and regularization parameters) appropriately. This is done with the
 # following function:
-
-max_iter = 30  # Maximum number of iterations of the DiffPIR algorithm
-
-# %%
+#
 # .. note::
 #
 #   We only use 30 steps to reduce the computational time of this example. As suggested by the authors of DiffPIR, the
-#   algorithm works best with ``max_iter = 100``.
+#   algorithm works best with ``diffusion_steps = 100``.
 #
 
+diffusion_steps = 30  # Maximum number of iterations of the DiffPIR algorithm
 
-def get_noise_schedule(max_iter=max_iter, num_train_timesteps=num_train_timesteps):
-    lambda_ = 7.0
+lambda_ = 7.0  # Regularization parameter
 
-    sigmas = []
-    sigma_ks = []
-    rhos = []
-    for i in range(num_train_timesteps):
-        sigmas.append(reduced_alpha_cumprod[num_train_timesteps - 1 - i])
-        sigma_ks.append((sqrt_1m_alphas_cumprod[i] / sqrt_alphas_cumprod[i]))
-        rhos.append(lambda_ * (sigma**2) / (sigma_ks[i] ** 2))
-    rhos, sigmas = torch.tensor(rhos).to(device), torch.tensor(sigmas).to(device)
+rhos = lambda_ * (sigma_noise**2) / (sigmas**2)
 
-    seq = np.sqrt(np.linspace(0, num_train_timesteps**2, max_iter))
-    seq = [int(s) for s in list(seq)]
-    seq[-1] = seq[-1] - 1
+# get timestep sequence
+seq = np.sqrt(np.linspace(0, T**2, diffusion_steps))
+seq = [int(s) for s in list(seq)]
+seq[-1] = seq[-1] - 1
 
-    return rhos, sigmas, seq
-
-
-rhos, sigmas, seq = get_noise_schedule()
 
 # Plot the noise and regularization schedules
 plt.figure(figsize=(6, 3))
+plt.rcParams.update({"font.size": 9})
 plt.subplot(121)
 plt.plot(
-    2 / rhos.numpy()[::-1]
+    2 / rhos.cpu().numpy()[::-1]
 )  # Note that the regularization parameter is 2/rho and not rho
 plt.xlabel(r"$t$")
 plt.ylabel(r"$\rho$")
 plt.subplot(122)
-plt.plot(sigmas.numpy()[::-1])
+plt.plot(sigmas.cpu().numpy()[::-1])
 plt.xlabel(r"$t$")
 plt.ylabel(r"$\sigma$")
 plt.suptitle("Regularisation parameter and noise schedules")
@@ -331,47 +271,58 @@ plt.show()
 # Initialization
 x = 2 * y - 1
 
-for i in range(len(seq)):
-    # Current noise level
-    curr_sigma = sigmas[seq[i]].cpu().numpy()
+with torch.no_grad():
+    for i in tqdm(range(len(seq))):
+        # Current and next noise levels
+        curr_sigma = sigmas[T - 1 - seq[i]].cpu().numpy()
 
-    # time step associated with the noise level sigmas[i]
-    t_i = find_nearest(reduced_alpha_cumprod, curr_sigma)
+        # 1. Denoising step
+        x0 = model(x, curr_sigma)
 
-    # Denoising step
-    noise_est_sample_var = model(x, torch.tensor([t_i]).to(y.device))
-    noise_est = noise_est_sample_var[:, :3, ...]
-    x0 = (
-        sqrt_recip_alphas_cumprod[t_i] * x
-        - sqrt_recipm1_alphas_cumprod[t_i] * noise_est
-    )
-    x0 = x0.clamp(-1, 1)
+        if not seq[i] == seq[-1]:
+            # 2. Data fidelity step
+            t_i = find_nearest(sigmas.cpu(), curr_sigma)
 
-    if not seq[i] == seq[-1]:
-        # Data fidelity step
-        x0_p = x0 / 2 + 0.5
-        x0_p = data_fidelity.prox(x0_p, y, physics, gamma=1 / (2 * rhos[t_i]))
-        x0 = x0_p * 2 - 1
+            x0 = data_fidelity.prox(x0, y, physics, gamma=1 / (2 * rhos[t_i]))
 
-        # Sampling step
-        t_im1 = find_nearest(
-            reduced_alpha_cumprod, sigmas[seq[i + 1]].cpu().numpy()
-        )  # time step associated with the next noise level
-        eps = (x - sqrt_alphas_cumprod[t_i] * x0) / sqrt_1m_alphas_cumprod[
-            t_i
-        ]  # effective noise
-        x = (
-            sqrt_alphas_cumprod[t_im1] * x0
-            + torch.sqrt(sqrt_1m_alphas_cumprod[t_im1] ** 2) * np.sqrt(1 - zeta) * eps
-            + sqrt_1m_alphas_cumprod[t_im1] * np.sqrt(zeta) * torch.randn_like(x)
-        )  # sampling
+            # Normalize data for sampling
+            x0 = 2 * x0 - 1
+            x = 2 * x - 1
+
+            # 3. Sampling step
+            next_sigma = sigmas[T - 1 - seq[i + 1]].cpu().numpy()
+            t_im1 = find_nearest(
+                sigmas, next_sigma
+            )  # time step associated with the next noise level
+
+            eps = (x - alphas[t_i].sqrt() * x0) / torch.sqrt(
+                1.0 - alphas[t_i]
+            )  # effective noise
+
+            x = alphas[t_im1].sqrt() * x0 + torch.sqrt(1.0 - alphas[t_im1]) * (
+                np.sqrt(1 - zeta) * eps + np.sqrt(zeta) * torch.randn_like(x)
+            )
+
+            # Rescale the output in [0, 1]
+            x = (x + 1) / 2
 
 
 # Plotting the results
-x = x / 2 + 0.5
-
 imgs = [y, x, x_true]
 plot(
     imgs,
-    titles=["measurement", "model output", "groundtruth"],
+    titles=["measurement", "model output", "ground-truth"],
 )
+
+
+# %%
+# Using the DiffPIR algorithm in your inverse problem
+# ---------------------------------------------
+# You can readily use this algorithm via the :meth:`deepinv.sampling.DiffPIR` class.
+#
+# ::
+#
+#       y = physics(x)
+#       model = dinv.sampling.DiffPIR(dinv.models.DiffUNet(), data_fidelity=dinv.optim.L2())
+#       xhat = model(y, physics)
+#
